requirements.txt
numpy
pandas
scipy
matplotlib
scikit-learn
xgboost
customtkinter
pyserial
